Troika	: fimc-is2: update cis driver
1. update 12a10/16885c code indets for fix SCAR
2. update 16885c setting to decrease vblank time
3. support LEC function for 12a10

Reviewed-on: https://gerrit.mot.com/1330699
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Submit-Approved: Jira Key

+++ drivers/media/platform/exynos/fimc-is2/sensor/module_framework/cis/fimc-is-cis-12a10ff.c
@@ -606,9 +606,9 @@ int sensor_12a10ff_cis_set_exposure_time(struct v4l2_subdev *subdev, struct ae_p
 	}
 
 	/* Short exposure */
-	short_coarse_val[0] = (short_coarse_int & 0xF000) >> 12;
-	short_coarse_val[1] = (short_coarse_int & 0x0FF0) >> 4;
-	short_coarse_val[2] = (short_coarse_int & 0x000F) << 4;
+	short_coarse_val[0] = (u8)((short_coarse_int & 0xF000) >> 12);
+	short_coarse_val[1] = (u8)((short_coarse_int & 0x0FF0) >> 4);
+	short_coarse_val[2] = (u8)((short_coarse_int & 0x000F) << 4);
 	ret = fimc_is_sensor_write8_array(client, 0x3500, short_coarse_val, 3);
 	if (ret < 0)
 		goto p_err;
@@ -1520,28 +1520,28 @@ int sensor_12a10ff_cis_wait_streamon(struct v4l2_subdev *subdev)
 
 	cis = (struct fimc_is_cis *)v4l2_get_subdevdata(subdev);
 	if (unlikely(!cis)) {
-	    err("cis is NULL");
-	    ret = -EINVAL;
-	    goto p_err;
+		err("cis is NULL");
+		ret = -EINVAL;
+		goto p_err;
 	}
 
 	cis_data = cis->cis_data;
 	if (unlikely(!cis_data)) {
-	    err("cis_data is NULL");
-	    ret = -EINVAL;
-	    goto p_err;
+		err("cis_data is NULL");
+		ret = -EINVAL;
+		goto p_err;
 	}
 
 	client = cis->client;
 	if (unlikely(!client)) {
-	    err("client is NULL");
-	    ret = -EINVAL;
-	    goto p_err;
+		err("client is NULL");
+		ret = -EINVAL;
+		goto p_err;
 	}
 
 	ret = fimc_is_sensor_read8(client, 0x483F, &sensor_fcount);
 	if (ret < 0)
-	    err("i2c transfer fail addr(%x), val(%x), ret = %d\n", 0x483F, sensor_fcount, ret);
+		err("i2c transfer fail addr(%x), val(%x), ret = %d\n", 0x483F, sensor_fcount, ret);
 
 	/*
 	 * Read sensor frame counter (sensor_fcount address = 0x0005)
@@ -1553,11 +1553,11 @@ int sensor_12a10ff_cis_wait_streamon(struct v4l2_subdev *subdev)
 
 		ret = fimc_is_sensor_read8(client, 0x483F, &sensor_fcount);
 		if (ret < 0)
-		    err("i2c transfer fail addr(%x), val(%x), ret = %d\n", 0x483F, sensor_fcount, ret);
+			err("i2c transfer fail addr(%x), val(%x), ret = %d\n", 0x483F, sensor_fcount, ret);
 
 		if (wait_cnt >= time_out_cnt) {
 			err("[MOD:D:%d] %s, Don't sensor stream on and time out, wait_limit(%d) > time_out(%d), sensor_fcount(%d)",
-				cis->id, __func__, wait_cnt, time_out_cnt, sensor_fcount);
+					cis->id, __func__, wait_cnt, time_out_cnt, sensor_fcount);
 			ret = -EINVAL;
 			goto p_err;
 		}
@@ -1577,49 +1577,49 @@ p_err:
 
 int sensor_12a10ff_cis_wait_streamoff(struct v4l2_subdev *subdev)
 {
-    int ret = 0;
-    struct fimc_is_cis *cis;
-    struct i2c_client *client;
-    cis_shared_data *cis_data;
-    u32 wait_cnt = 0, time_out_cnt = 250;
-    u8 sensor_fcount = 0;
+	int ret = 0;
+	struct fimc_is_cis *cis;
+	struct i2c_client *client;
+	cis_shared_data *cis_data;
+	u32 wait_cnt = 0, time_out_cnt = 250;
+	u8 sensor_fcount = 0;
 
-    BUG_ON(!subdev);
+	BUG_ON(!subdev);
 
-    cis = (struct fimc_is_cis *)v4l2_get_subdevdata(subdev);
-    cis_data = cis->cis_data;
-    client = cis->client;
+	cis = (struct fimc_is_cis *)v4l2_get_subdevdata(subdev);
+	cis_data = cis->cis_data;
+	client = cis->client;
 
-    I2C_MUTEX_LOCK(cis->i2c_lock);
-    ret = fimc_is_sensor_read8(client, 0x483F, &sensor_fcount);
-    I2C_MUTEX_UNLOCK(cis->i2c_lock);
+	I2C_MUTEX_LOCK(cis->i2c_lock);
+	ret = fimc_is_sensor_read8(client, 0x483F, &sensor_fcount);
+	I2C_MUTEX_UNLOCK(cis->i2c_lock);
 
-    if (ret < 0)
-            err("i2c transfer fail addr(%x), val(%x), ret = %d\n", 0x483F, sensor_fcount, ret);
+	if (ret < 0)
+		err("i2c transfer fail addr(%x), val(%x), ret = %d\n", 0x483F, sensor_fcount, ret);
 
-    while (sensor_fcount != 0x00) {
-            I2C_MUTEX_LOCK(cis->i2c_lock);
-            ret = fimc_is_sensor_read8(client, 0x483F, &sensor_fcount);
-            I2C_MUTEX_UNLOCK(cis->i2c_lock);
-            if (ret < 0)
-                    err("i2c transfer fail addr(%x), val(%x), ret = %d\n", 0x483F, sensor_fcount, ret);
+	while (sensor_fcount != 0x00) {
+		I2C_MUTEX_LOCK(cis->i2c_lock);
+		ret = fimc_is_sensor_read8(client, 0x483F, &sensor_fcount);
+		I2C_MUTEX_UNLOCK(cis->i2c_lock);
+		if (ret < 0)
+			err("i2c transfer fail addr(%x), val(%x), ret = %d\n", 0x483F, sensor_fcount, ret);
 
-			usleep_range(CIS_STREAM_OFF_WAIT_TIME, CIS_STREAM_OFF_WAIT_TIME);
-            wait_cnt++;
+		usleep_range(CIS_STREAM_OFF_WAIT_TIME, CIS_STREAM_OFF_WAIT_TIME);
+		wait_cnt++;
 
-            if (wait_cnt >= time_out_cnt) {
-                    err("[MOD:D:%d] %s, time out, wait_limit(%d) > time_out(%d), sensor_fcount(%d)",
-                                    cis->id, __func__, wait_cnt, time_out_cnt, sensor_fcount);
-                    ret = -EINVAL;
-                    goto p_err;
-            }
+		if (wait_cnt >= time_out_cnt) {
+			err("[MOD:D:%d] %s, time out, wait_limit(%d) > time_out(%d), sensor_fcount(%d)",
+					cis->id, __func__, wait_cnt, time_out_cnt, sensor_fcount);
+			ret = -EINVAL;
+			goto p_err;
+		}
 
-            dbg_sensor(1, "[MOD:D:%d] %s, sensor_fcount(%d), (wait_limit(%d) < time_out(%d))\n",
-                            cis->id, __func__, sensor_fcount, wait_cnt, time_out_cnt);
-    }
+		dbg_sensor(1, "[MOD:D:%d] %s, sensor_fcount(%d), (wait_limit(%d) < time_out(%d))\n",
+				cis->id, __func__, sensor_fcount, wait_cnt, time_out_cnt);
+	}
 
 p_err:
-    return ret;
+	return ret;
 }
 
 int sensor_12a10ff_cis_long_term_exposure(struct v4l2_subdev *subdev)
